bitbake.conf/sanity: Drop makeinfo as being required on buildhost
This is a long standing 'odd' dependency which we've meant to
resolve. We shouldn't need it and it introduces reproducability issues.

We already have texinfo-dummy-native and texinfo-native which can
provide it but the work to remove the hosttool was never completed.

After cleaning up texinfo.bbclass, this can now be removed with
minimal impact on build time.

[YOCTO #13753]

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>